requirements.txt
asgiref==3.7.2
Django==4.2.10
pillow==10.2.0
psycopg2==2.9.9
sqlparse==0.4.4
typing_extensions==4.10.0
phonenumbers==8.13.30
django-phonenumber-field[phonenumbers]
